Removing fwerror package and moving FunctionError to function package
