+++ function/arguments_data_test.go
@@ -11,7 +11,6 @@ import (
 
 	"github.com/hashicorp/terraform-plugin-framework/attr"
 	"github.com/hashicorp/terraform-plugin-framework/function"
-	"github.com/hashicorp/terraform-plugin-framework/fwerror"
 	"github.com/hashicorp/terraform-plugin-framework/types/basetypes"
 )
 
@@ -89,14 +88,14 @@ func TestArgumentsDataGet(t *testing.T) {
 		argumentsData function.ArgumentsData
 		targets       []any
 		expected      []any
-		expectedErr   fwerror.FunctionErrors
+		expectedErr   function.FunctionErrors
 	}{
 		"no-argument-data": {
 			argumentsData: function.NewArgumentsData(nil),
 			targets:       []any{new(bool)},
 			expected:      []any{new(bool)},
-			expectedErr: fwerror.FunctionErrors{
-				fwerror.NewFunctionError("Invalid Argument Data Usage: When attempting to fetch argument data during the function call, the provider code incorrectly attempted to read argument data. " +
+			expectedErr: function.FunctionErrors{
+				function.NewFunctionError("Invalid Argument Data Usage: When attempting to fetch argument data during the function call, the provider code incorrectly attempted to read argument data. " +
 					"This is always an issue in the provider code and should be reported to the provider developers.\n\n" +
 					"Function does not have argument data."),
 			},
@@ -108,8 +107,8 @@ func TestArgumentsDataGet(t *testing.T) {
 			}),
 			targets:  []any{new(bool)},
 			expected: []any{new(bool)},
-			expectedErr: fwerror.FunctionErrors{
-				fwerror.NewFunctionError("Invalid Argument Data Usage: When attempting to fetch argument data during the function call, the provider code incorrectly attempted to read argument data. " +
+			expectedErr: function.FunctionErrors{
+				function.NewFunctionError("Invalid Argument Data Usage: When attempting to fetch argument data during the function call, the provider code incorrectly attempted to read argument data. " +
 					"The Get call requires all parameters and the final variadic parameter, if implemented, to be in the targets. " +
 					"This is always an error in the provider code and should be reported to the provider developers.\n\n" +
 					"Given targets count: 1, expected targets count: 2"),
@@ -121,8 +120,8 @@ func TestArgumentsDataGet(t *testing.T) {
 			}),
 			targets:  []any{new(bool), new(bool)},
 			expected: []any{new(bool), new(bool)},
-			expectedErr: fwerror.FunctionErrors{
-				fwerror.NewFunctionError("Invalid Argument Data Usage: When attempting to fetch argument data during the function call, the provider code incorrectly attempted to read argument data. " +
+			expectedErr: function.FunctionErrors{
+				function.NewFunctionError("Invalid Argument Data Usage: When attempting to fetch argument data during the function call, the provider code incorrectly attempted to read argument data. " +
 					"The Get call requires all parameters and the final variadic parameter, if implemented, to be in the targets. " +
 					"This is always an error in the provider code and should be reported to the provider developers.\n\n" +
 					"Given targets count: 2, expected targets count: 1"),
@@ -134,8 +133,8 @@ func TestArgumentsDataGet(t *testing.T) {
 			}),
 			targets:  []any{new(basetypes.StringValue)},
 			expected: []any{new(basetypes.StringValue)},
-			expectedErr: fwerror.FunctionErrors{
-				fwerror.NewFunctionError("Value Conversion Error: An unexpected error was encountered trying to convert into a Terraform value. " +
+			expectedErr: function.FunctionErrors{
+				function.NewFunctionError("Value Conversion Error: An unexpected error was encountered trying to convert into a Terraform value. " +
 					"This is always an error in the provider. Please report the following to the provider developer:\n\n" +
 					"Cannot use attr.Value basetypes.StringValue, only basetypes.BoolValue is supported because basetypes.BoolType is the type in the schema"),
 			},
@@ -230,15 +229,15 @@ func TestArgumentsDataGetArgument(t *testing.T) {
 		position      int
 		target        any
 		expected      any
-		expectedErr   fwerror.FunctionErrors
+		expectedErr   function.FunctionErrors
 	}{
 		"no-argument-data": {
 			argumentsData: function.NewArgumentsData(nil),
 			position:      0,
 			target:        new(bool),
 			expected:      new(bool),
-			expectedErr: fwerror.FunctionErrors{
-				fwerror.NewArgumentFunctionError(0, "Invalid Argument Data Usage: When attempting to fetch argument data during the function call, the provider code incorrectly attempted to read argument data. "+
+			expectedErr: function.FunctionErrors{
+				function.NewArgumentFunctionError(0, "Invalid Argument Data Usage: When attempting to fetch argument data during the function call, the provider code incorrectly attempted to read argument data. "+
 					"This is always an issue in the provider code and should be reported to the provider developers.\n\n"+
 					"Function does not have argument data."),
 			},
@@ -250,8 +249,8 @@ func TestArgumentsDataGetArgument(t *testing.T) {
 			position: 1,
 			target:   new(bool),
 			expected: new(bool),
-			expectedErr: fwerror.FunctionErrors{
-				fwerror.NewArgumentFunctionError(1, "Invalid Argument Data Position: When attempting to fetch argument data during the function call, the provider code attempted to read a non-existent argument position. "+
+			expectedErr: function.FunctionErrors{
+				function.NewArgumentFunctionError(1, "Invalid Argument Data Position: When attempting to fetch argument data during the function call, the provider code attempted to read a non-existent argument position. "+
 					"Function argument positions are 0-based and any final variadic parameter is represented as one argument position with an ordered list of the parameter data type. "+
 					"This is always an error in the provider code and should be reported to the provider developers.\n\n"+
 					"Given argument position: 1, last argument position: 0"),
@@ -264,8 +263,8 @@ func TestArgumentsDataGetArgument(t *testing.T) {
 			position: 0,
 			target:   new(basetypes.StringValue),
 			expected: new(basetypes.StringValue),
-			expectedErr: fwerror.FunctionErrors{
-				fwerror.NewFunctionError("Value Conversion Error: An unexpected error was encountered trying to convert into a Terraform value. " +
+			expectedErr: function.FunctionErrors{
+				function.NewFunctionError("Value Conversion Error: An unexpected error was encountered trying to convert into a Terraform value. " +
 					"This is always an error in the provider. Please report the following to the provider developer:\n\n" +
 					"Cannot use attr.Value basetypes.StringValue, only basetypes.BoolValue is supported because basetypes.BoolType is the type in the schema"),
 			},
