+++ function/function_error_basic.go
@@ -1,7 +1,7 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
 
-package fwerror
+package function
 
 import (
 	"errors"
