+++ function/number_return.go
@@ -7,7 +7,6 @@ import (
 	"context"
 
 	"github.com/hashicorp/terraform-plugin-framework/attr"
-	"github.com/hashicorp/terraform-plugin-framework/fwerror"
 	"github.com/hashicorp/terraform-plugin-framework/types/basetypes"
 )
 
@@ -39,8 +38,8 @@ func (r NumberReturn) GetType() attr.Type {
 }
 
 // NewResultData returns a new result data based on the type.
-func (r NumberReturn) NewResultData(ctx context.Context) (ResultData, fwerror.FunctionErrors) {
-	var funcErrs fwerror.FunctionErrors
+func (r NumberReturn) NewResultData(ctx context.Context) (ResultData, FunctionErrors) {
+	var funcErrs FunctionErrors
 
 	value := basetypes.NewNumberUnknown()
 
@@ -50,7 +49,7 @@ func (r NumberReturn) NewResultData(ctx context.Context) (ResultData, fwerror.Fu
 
 	valuable, diags := r.CustomType.ValueFromNumber(ctx, value)
 
-	funcErrs.Append(fwerror.FunctionErrorsFromDiags(ctx, diags)...)
+	funcErrs.Append(FunctionErrorsFromDiags(ctx, diags)...)
 
 	return NewResultData(valuable), funcErrs
 }
