+++ function/object_return.go
@@ -7,7 +7,6 @@ import (
 	"context"
 
 	"github.com/hashicorp/terraform-plugin-framework/attr"
-	"github.com/hashicorp/terraform-plugin-framework/fwerror"
 	"github.com/hashicorp/terraform-plugin-framework/types/basetypes"
 )
 
@@ -42,8 +41,8 @@ func (r ObjectReturn) GetType() attr.Type {
 }
 
 // NewResultData returns a new result data based on the type.
-func (r ObjectReturn) NewResultData(ctx context.Context) (ResultData, fwerror.FunctionErrors) {
-	var funcErrs fwerror.FunctionErrors
+func (r ObjectReturn) NewResultData(ctx context.Context) (ResultData, FunctionErrors) {
+	var funcErrs FunctionErrors
 
 	value := basetypes.NewObjectUnknown(r.AttributeTypes)
 
@@ -53,7 +52,7 @@ func (r ObjectReturn) NewResultData(ctx context.Context) (ResultData, fwerror.Fu
 
 	valuable, diags := r.CustomType.ValueFromObject(ctx, value)
 
-	funcErrs.Append(fwerror.FunctionErrorsFromDiags(ctx, diags)...)
+	funcErrs.Append(FunctionErrorsFromDiags(ctx, diags)...)
 
 	return NewResultData(valuable), funcErrs
 }
