+++ function/result_data_test.go
@@ -11,7 +11,6 @@ import (
 
 	"github.com/hashicorp/terraform-plugin-framework/attr"
 	"github.com/hashicorp/terraform-plugin-framework/function"
-	"github.com/hashicorp/terraform-plugin-framework/fwerror"
 	"github.com/hashicorp/terraform-plugin-framework/types/basetypes"
 )
 
@@ -22,14 +21,14 @@ func TestResultDataSet(t *testing.T) {
 		resultData  function.ResultData
 		value       any
 		expected    attr.Value
-		expectedErr fwerror.FunctionErrors
+		expectedErr function.FunctionErrors
 	}{
 		"nil": {
 			resultData: function.NewResultData(basetypes.NewBoolUnknown()),
 			value:      nil,
 			expected:   basetypes.NewBoolUnknown(),
-			expectedErr: fwerror.FunctionErrors{
-				fwerror.NewFunctionError("Value Conversion Error: An unexpected error was encountered trying to convert from value. " +
+			expectedErr: function.FunctionErrors{
+				function.NewFunctionError("Value Conversion Error: An unexpected error was encountered trying to convert from value. " +
 					"This is always an error in the provider. Please report the following to the provider developer:\n\n" +
 					"cannot construct attr.Type from <nil> (invalid)"),
 			},
@@ -38,8 +37,8 @@ func TestResultDataSet(t *testing.T) {
 			resultData: function.NewResultData(basetypes.NewBoolUnknown()),
 			value:      basetypes.NewStringValue("test"),
 			expected:   basetypes.NewBoolUnknown(),
-			expectedErr: fwerror.FunctionErrors{
-				fwerror.NewFunctionError("Value Conversion Error: An unexpected error was encountered while verifying an attribute value matched its expected type to prevent unexpected behavior or panics. " +
+			expectedErr: function.FunctionErrors{
+				function.NewFunctionError("Value Conversion Error: An unexpected error was encountered while verifying an attribute value matched its expected type to prevent unexpected behavior or panics. " +
 					"This is always an error in the provider. Please report the following to the provider developer:\n\n" +
 					"Expected framework type from provider logic: basetypes.BoolType / underlying type: tftypes.Bool\n" +
 					"Received framework type from provider logic: basetypes.StringType / underlying type: tftypes.String\n" +
