+++ function/return.go
@@ -7,7 +7,6 @@ import (
 	"context"
 
 	"github.com/hashicorp/terraform-plugin-framework/attr"
-	"github.com/hashicorp/terraform-plugin-framework/fwerror"
 )
 
 // Return is the interface for defining function return data.
@@ -22,5 +21,5 @@ type Return interface {
 	// best approximation of an invalid value) of the corresponding data type.
 	// The Function type Run method is expected to overwrite the value before
 	// returning.
-	NewResultData(context.Context) (ResultData, fwerror.FunctionErrors)
+	NewResultData(context.Context) (ResultData, FunctionErrors)
 }
