+++ function/set_return.go
@@ -7,7 +7,6 @@ import (
 	"context"
 
 	"github.com/hashicorp/terraform-plugin-framework/attr"
-	"github.com/hashicorp/terraform-plugin-framework/fwerror"
 	"github.com/hashicorp/terraform-plugin-framework/types/basetypes"
 )
 
@@ -46,8 +45,8 @@ func (r SetReturn) GetType() attr.Type {
 }
 
 // NewResultData returns a new result data based on the type.
-func (r SetReturn) NewResultData(ctx context.Context) (ResultData, fwerror.FunctionErrors) {
-	var funcErrs fwerror.FunctionErrors
+func (r SetReturn) NewResultData(ctx context.Context) (ResultData, FunctionErrors) {
+	var funcErrs FunctionErrors
 
 	value := basetypes.NewSetUnknown(r.ElementType)
 
@@ -57,7 +56,7 @@ func (r SetReturn) NewResultData(ctx context.Context) (ResultData, fwerror.Funct
 
 	valuable, diags := r.CustomType.ValueFromSet(ctx, value)
 
-	funcErrs.Append(fwerror.FunctionErrorsFromDiags(ctx, diags)...)
+	funcErrs.Append(FunctionErrorsFromDiags(ctx, diags)...)
 
 	return NewResultData(valuable), funcErrs
 }
