--- fwerror/doc.go
@@ -1,6 +1,0 @@
-// Copyright (c) HashiCorp, Inc.
-// SPDX-License-Identifier: MPL-2.0
-
-// Package fwerror contains type and value interfaces for core framework and
-// provider-defined errors.
-package fwerror
